api/requirements/base.txt
Django==5.2.6
djangorestframework==3.16.1
django-filter
djangorestframework-simplejwt
celery
